Revert "Rename agent/proxy package to reflect that it is limited to managed proxy processes. (#4550)"
This reverts commit e7e7c7d5216491ea581851bea486c3bb8e4f8956.
