+++ agent/proxy/daemon_test.go
@@ -1,4 +1,4 @@
-package proxyprocess
+package proxy
 
 import (
 	"io/ioutil"
