+++ agent/proxy/manager_test.go
@@ -1,4 +1,4 @@
-package proxyprocess
+package proxy
 
 import (
 	"io/ioutil"
