+++ agent/proxy/process_windows.go
@@ -1,6 +1,6 @@
 // +build windows
 
-package proxyprocess
+package proxy
 
 import (
 	"os"
