+++ agent/proxy/root.go
@@ -1,4 +1,4 @@
-package proxyprocess
+package proxy
 
 import (
 	"os"
